Refactoring, updated animations, throwing knives
Moved code from the game loop of main.py to a separate module named game_functions.py to reduce clutter in main. Updated the skeleton animation and remapped the sprites for it and haunted axe. Implemented a throwing knife (press g). Limited to 3 on screen at a time and only travels to the right currently. Need to implement leftward shot and allow it to damage enemies.